Move skabase.utils files up one directory
This simplifies the naming when importing, which "feels" a little
nicer.

JIRA:  LMC-428
